chore(examples): improved tests-cases package and added some reusable functions
