+++ packages/vm/examples/run-deployment-transaction.ts
@@ -1,0 +1,23 @@
+import {toBuffer} from 'ethereumjs-util'
+import VM from "@ethereumjs/vm";
+import {debugLog, deploySimpleStorage, execute, insertOne, privateKey} from "@ethereumjs/test-cases";
+
+const main = async () => {
+  const vm = new VM()
+
+  // used to sign transactions and generate addresses
+  const privateKeyBuf: Buffer = toBuffer(privateKey)
+  const {address} = await insertOne(vm)
+
+  debugLog(`Account ${address.toString()} has been saved into local chain`)
+
+  const deployContractTransaction = deploySimpleStorage()
+  await execute(vm, deployContractTransaction, privateKeyBuf)
+}
+
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error)
+    process.exit(1)
+  })
