+++ packages/vm/package.json
@@ -44,7 +44,7 @@
     "lint:fix": "../../config/cli/lint-fix.sh",
     "formatTest": "node ./scripts/formatTest",
     "tsc": "../../config/cli/ts-compile.sh",
-    "examples": "ts-node ./examples/run-transactions-complete.ts"
+    "examples": "ts-node examples/*.ts"
   },
   "dependencies": {
     "@ethereumjs/block": "^3.6.0",
